Change card direction to horizontally position Add the StudentsCards class to separate card creation funcionality from main program class
